Move event/notification/suggestions shaping out; rewards service
- responders.event: materialize_event + notification_schema_from_orm
- responders.suggestions: submission_suggestion + answer_suggest_edit
- responders.user: user_schema_from_orm
- services.rewards + thin rewards endpoint
- Materializer methods are thin delegates; materialize.py ~429 LOC
