+++ chafan_core/app/api/api_v1/endpoints/rewards.py
@@ -1,27 +1,12 @@
-import datetime
 from typing import Any, List
 
 from fastapi import APIRouter, Depends
-from pydantic.tools import parse_obj_as
-from sqlalchemy.orm import Session
 
-from chafan_core.app import crud, rep_manager, schemas
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.common import OperationType
-from chafan_core.app.materialize import can_read_answer
-from chafan_core.app.user_permission import user_in_site
-from chafan_core.app.schemas.event import (
-    ClaimAnswerQuestionRewardInternal,
-    CreateAnswerQuestionRewardInternal,
-    EventInternal,
-)
-from chafan_core.app.schemas.reward import (
-    AnsweredQuestionCondition,
-    RewardCondition,
-    RewardCreate,
-)
-from chafan_core.utils.base import HTTPException_
+from chafan_core.app.schemas.reward import RewardCreate
+from chafan_core.app.services import rewards as rewards_service
 
 router = APIRouter()
 
@@ -31,11 +16,7 @@ router = APIRouter()
 def get_rewards(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
 ) -> Any:
-    current_user = ctx.get_current_active_user()
-    received = current_user.incoming_rewards
-    given = current_user.outgoing_rewards
-    rewards = sorted(received + given, key=lambda r: r.created_at, reverse=True)
-    return [ctx.materializer.reward_schema_from_orm(r) for r in rewards]
+    return rewards_service.list_rewards(ctx)
 
 
 @router.post("/", response_model=schemas.Reward)
@@ -44,54 +25,7 @@ def create_reward(
     *,
     reward_in: RewardCreate,
 ) -> Any:
-    db = ctx.get_db()
-    current_user = ctx.get_current_active_user()
-    receiver = crud.user.get_by_uuid(db, uuid=reward_in.receiver_uuid)
-    if receiver is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The receiver doesn't exist in the system.",
-        )
-    if reward_in.condition:
-        if isinstance(reward_in.condition.content, AnsweredQuestionCondition):
-            question = crud.question.get_by_uuid(
-                db, uuid=reward_in.condition.content.question_uuid
-            )
-            assert question is not None
-            if not user_in_site(
-                db,
-                site=question.site,
-                user_id=receiver.id,
-                op_type=OperationType.WriteSiteAnswer,
-            ):
-                raise HTTPException_(
-                    status_code=400,
-                    detail="The receiver can't post answer for that question.",
-                )
-    if current_user.remaining_coins < reward_in.coin_amount:
-        raise HTTPException_(
-            status_code=400,
-            detail="Insufficient coins.",
-        )
-    reward = crud.reward.create_with_giver(db, obj_in=reward_in, giver=current_user)
-    # FIXME: reward event for other types of rewards (including non-conditonal ones)
-    reward_event = None
-    if reward_in.condition:
-        if isinstance(reward_in.condition.content, AnsweredQuestionCondition):
-            reward_event = CreateAnswerQuestionRewardInternal(
-                subject_id=current_user.id,
-                reward_id=reward.id,
-            )
-    if reward_event is not None:
-        crud.notification.create_with_content(
-            ctx,
-            receiver_id=receiver.id,
-            event=EventInternal(
-                created_at=reward.created_at,
-                content=reward_event,
-            ),
-        )
-    return ctx.materializer.reward_schema_from_orm(reward)
+    return rewards_service.create_reward(ctx, reward_in=reward_in)
 
 
 @router.post("/{id}/claim", response_model=schemas.Reward)
@@ -100,112 +34,13 @@ def claim_reward(
     *,
     id: int,
 ) -> Any:
-    db = ctx.get_db()
-    reward = crud.reward.get(db, id=id)
-    current_user = ctx.get_current_active_user()
-    if reward is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward doesn't exist in the system.",
-        )
-    if reward.receiver_id != current_user.id:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is not for current user",
-        )
-    if reward.claimed_at is not None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is already claimed",
-        )
-    if reward.refunded_at is not None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is already refunded",
-        )
-    utc_now = datetime.datetime.now(tz=datetime.timezone.utc)
-    if reward.expired_at < utc_now:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is already expired",
-        )
-    claimable = False
-    if reward.condition is not None:
-        condition = parse_obj_as(RewardCondition, reward.condition)
-        if isinstance(condition.content, AnsweredQuestionCondition):
-            question = crud.question.get_by_uuid(
-                db, uuid=condition.content.question_uuid
-            )
-            assert question is not None
-            claimable = any(
-                can_read_answer(db, answer=answer, principal_id=reward.giver_id)
-                for answer in question.answers
-            )
-    else:
-        claimable = True
-    if not claimable:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward condition is not met yet",
-        )
-    reward.claimed_at = utc_now
-    rep_manager.award_coins(db, current_user, reward.coin_amount, "reward_claim")
-    db.add(reward)
-    db.commit()
-    db.refresh(reward)
-    reward_data = ctx.materializer.reward_schema_from_orm(reward)
-    reward_event = None
-    if reward_data.condition:
-        if isinstance(reward_data.condition.content, AnsweredQuestionCondition):
-            reward_event = ClaimAnswerQuestionRewardInternal(
-                subject_id=current_user.id,
-                reward_id=reward.id,
-            )
-    if reward_event is not None:
-        crud.notification.create_with_content(
-            ctx,
-            receiver_id=reward.giver.id,
-            event=EventInternal(
-                created_at=utc_now,
-                content=reward_event,
-            ),
-        )
-    return reward_data
+    return rewards_service.claim_reward(ctx, reward_id=id)
 
 
 @router.post("/{id}/refund", response_model=schemas.Reward)
 def refund_reward(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     *,
-    db: Session = Depends(deps.get_db),
     id: int,
 ) -> Any:
-    current_user = ctx.get_current_active_user()
-    reward = crud.reward.get(db, id=id)
-    if reward is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward doesn't exist in the system.",
-        )
-    if reward.giver_id != current_user.id:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is not for current user",
-        )
-    if reward.claimed_at is not None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is already claimed",
-        )
-    if reward.refunded_at is not None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The reward is already refunded",
-        )
-    utc_now = datetime.datetime.now(tz=datetime.timezone.utc)
-    reward.refunded_at = utc_now
-    rep_manager.award_coins(db, current_user, reward.coin_amount, "reward_refund")
-    db.add(reward)
-    db.commit()
-    db.refresh(reward)
-    return ctx.materializer.reward_schema_from_orm(reward)
+    return rewards_service.refund_reward(ctx, reward_id=id)
