+++ chafan_core/app/responders/__init__.py
@@ -1,4 +1,16 @@
-from . import question, submission, article, answer, site, rss, comment, misc
+from . import (
+    question,
+    submission,
+    article,
+    answer,
+    site,
+    rss,
+    comment,
+    misc,
+    event,
+    suggestions,
+    user,
+)
 
 # responders 的设计目的是 models (crud) -> schema (api endpoint)
 #   1. responders 不需要鉴权
